Move storybook helpers to ui directory
