+++ src/ui/nav/navbar/buttons/toggle_menu_button/stories/toggle_menu_button.stories.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/react';
 import React, { useState } from 'react';
 
-import { StoryGroup } from '../../../../../../util/storybook/stories';
+import { StoryGroup } from '../../../../../storybook/stories';
 import { ToggleMenuButton } from '../toggle_menu_button';
 
 storiesOf('UI/Nav/Navbar/Buttons', module)
