+++ src/ui/text/stories/text.stories.tsx
@@ -2,7 +2,7 @@ import { boolean, select } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import React from 'react';
 
-import { StoryGroup } from '../../../util/storybook/stories';
+import { StoryGroup } from '../../storybook/stories';
 import { Text } from '../text';
 
 storiesOf('UI/Text', module)
